requirements.txt
opencv-python
mediapipe
tensorflow
screeninfo
pandas
# For macOS mouse control
pyobjc-core
pyobjc-framework-quartz